pyproject.toml
[project]
name = "french-admin-agent"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.2.0",
    "qdrant-client>=1.9.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "ragas>=0.1.7",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "datasets",
    "langchain-huggingface",
    "langchain-openai",
]
